完成支付功能，但是全局事务并没有添加

+++ bank9001/src/main/java/com/nnxy/bank/entity/AccountEntity.java
@@ -36,7 +36,7 @@ public class AccountEntity implements Serializable {
 	/**
 	 * 账户余额
 	 */
-	private Long aMoney;
+	private Double aMoney;
 
 	public static long getSerialVersionUID() {
 		return serialVersionUID;
@@ -54,7 +54,7 @@ public class AccountEntity implements Serializable {
 		return aPassword;
 	}
 
-	public Long getaMoney() {
+	public Double getaMoney() {
 		return aMoney;
 	}
 
@@ -70,7 +70,7 @@ public class AccountEntity implements Serializable {
 		this.aPassword = aPassword;
 	}
 
-	public void setaMoney(Long aMoney) {
+	public void setaMoney(Double aMoney) {
 		this.aMoney = aMoney;
 	}
 }
