+++ bank9001/src/main/java/com/nnxy/bank/entity/FlowEntity.java
@@ -27,7 +27,7 @@ public class FlowEntity implements Serializable {
 	/**
 	 * 金额
 	 */
-	private Long fMoney;
+	private Double fMoney;
 	/**
 	 * 类型
 	 */
@@ -53,11 +53,11 @@ public class FlowEntity implements Serializable {
 		this.fId = fId;
 	}
 
-	public Long getfMoney() {
+	public Double getfMoney() {
 		return fMoney;
 	}
 
-	public void setfMoney(Long fMoney) {
+	public void setfMoney(Double fMoney) {
 		this.fMoney = fMoney;
 	}
